+++ Libraries/Editor/TextEditor.hpp
@@ -318,7 +318,7 @@ public:
   String GetText(int start, int end);
   int GetRuneAt(int position);
   //Event Processing
-  int SendEditor(unsigned int Msg, u64 wParam = 0, s64 lParam = 0);
+  intptr_t SendEditor(unsigned int Msg, u64 wParam = 0, s64 lParam = 0);
   void SetAStyle(int style, ByteColor fore, ByteColor back, int size = 0, cstr face = 0);
   //Send to derived classes
   virtual void OnNotify(Scintilla::SCNotification& scn);
